requirements.txt
alembic==1.7.5
click==8.0.3
feedparser==6.0.8
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
psycopg2==2.9.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
sgmllib3k==1.0.0
six==1.16.0
SQLAlchemy==1.4.27
Werkzeug==2.0.2
